[client][#8]순환참조 해결을 위해 naverMap을 자식 모듈로 빼냄
